# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Clearbook API. Issuer-aware brokerage statements for tokenized stocks on Solana.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallets/{address}/index:
    post:
      operationId: indexWallet
      tags: [wallets]
      summary: Index or refresh a wallet
      description: Discovers token accounts, fetches history and rebuilds the ledger. Demo wallets load instantly.
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
      responses:
        "200":
          description: Index status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/UpstreamError"
  /wallets/{address}/reset:
    post:
      operationId: resetWallet
      tags: [wallets]
      summary: Remove simulated trades and cached ledger for a wallet
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
      responses:
        "200":
          description: Status after reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
  /wallets/{address}/status:
    get:
      operationId: getWalletStatus
      tags: [wallets]
      summary: Indexing status for a wallet
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
  /wallets/{address}/portfolio:
    get:
      operationId: getPortfolio
      tags: [wallets]
      summary: Portfolio valuation, positions and totals
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
        - $ref: "#/components/parameters/CostMethod"
      responses:
        "200":
          description: Portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /wallets/{address}/lots:
    get:
      operationId: listLots
      tags: [wallets]
      summary: Tax lots for a wallet
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
        - $ref: "#/components/parameters/CostMethod"
        - name: mint
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/LotStatusFilter"
      responses:
        "200":
          description: Lots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lot"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /wallets/{address}/activity:
    get:
      operationId: listActivity
      tags: [wallets]
      summary: Ledger activity for a wallet
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
        - $ref: "#/components/parameters/CostMethod"
        - name: mint
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - name: cursor
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Activity page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /wallets/{address}/events:
    get:
      operationId: listCorporateActions
      tags: [wallets]
      summary: Corporate action and multiplier events for a wallet
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CorporateActionEvent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /wallets/{address}/pricing:
    get:
      operationId: getPricingStatus
      tags: [wallets]
      summary: Pricing source status for a wallet's holdings
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
      responses:
        "200":
          description: Pricing status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /wallets/{address}/statements:
    get:
      operationId: listStatements
      tags: [statements]
      summary: Statements generated for a wallet
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
      responses:
        "200":
          description: Statements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatementSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      operationId: createStatement
      tags: [statements]
      summary: Generate a statement for a period
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatementRequest"
      responses:
        "201":
          description: Statement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Statement"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /statements/{statementId}:
    get:
      operationId: getStatement
      tags: [statements]
      summary: Fetch a generated statement
      parameters:
        - $ref: "#/components/parameters/StatementId"
      responses:
        "200":
          description: Statement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Statement"
        "404":
          $ref: "#/components/responses/NotFound"
  /statements/{statementId}/export.csv:
    get:
      operationId: exportStatementCsv
      tags: [statements]
      summary: Download the statement as CSV
      parameters:
        - $ref: "#/components/parameters/StatementId"
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/NotFound"
  /statements/{statementId}/export.pdf:
    get:
      operationId: exportStatementPdf
      tags: [statements]
      summary: Download the statement as PDF
      parameters:
        - $ref: "#/components/parameters/StatementId"
      responses:
        "200":
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
  /statements/{statementId}/notarization:
    get:
      operationId: prepareNotarization
      tags: [statements]
      summary: Memo payload the wallet signs to notarize the statement
      parameters:
        - $ref: "#/components/parameters/StatementId"
        - name: payer
          in: query
          required: false
          description: Wallet public key that will pay for the memo transaction. When present the response includes an unsigned transaction.
          schema:
            type: string
      responses:
        "200":
          description: Notarization payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotarizationPayload"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: submitNotarization
      tags: [statements]
      summary: Record a signed memo transaction or create a simulated proof
      parameters:
        - $ref: "#/components/parameters/StatementId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotarizationSubmission"
      responses:
        "200":
          description: Proof
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatementProof"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/UpstreamError"
  /wallets/{address}/trades/quote:
    post:
      operationId: quoteTrade
      tags: [trades]
      summary: Quote a sell of a position and preview the lots it closes
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeQuoteRequest"
      responses:
        "200":
          description: Quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeQuote"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/UpstreamError"
  /wallets/{address}/trades/prepare:
    post:
      operationId: prepareTrade
      tags: [trades]
      summary: Build the Jupiter swap transaction for a quote
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradePrepareRequest"
      responses:
        "200":
          description: Unsigned transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeTransaction"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/UpstreamError"
  /wallets/{address}/trades/confirm:
    post:
      operationId: confirmTrade
      tags: [trades]
      summary: Confirm a signed swap on chain and post it to the ledger
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeConfirmRequest"
      responses:
        "200":
          description: Trade result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/UpstreamError"
  /wallets/{address}/trades/simulate:
    post:
      operationId: simulateTrade
      tags: [trades]
      summary: Post a simulated sell to the ledger at the quoted price
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeSimulateRequest"
      responses:
        "200":
          description: Trade result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    WalletAddress:
      name: address
      in: path
      required: true
      description: Base58 Solana wallet address or a demo wallet id
      schema:
        type: string
    StatementId:
      name: statementId
      in: path
      required: true
      schema:
        type: string
    CostMethod:
      name: method
      in: query
      required: false
      description: Lot relief method. FIFO is the default.
      schema:
        $ref: "#/components/schemas/CostMethod"
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    UpstreamError:
      description: An upstream data source failed or rejected the request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    LotStatusFilter:
      type: string
      enum: [open, closed, all]
...
    WalletStatus:
      type: object
      properties:
        address:
          type: string
        displayAddress:
          type: string
        isDemo:
          type: boolean
        demoLabel:
          type: string
          nullable: true
        state:
          type: string
          enum: [not_indexed, indexing, ready, partial, empty, error]
        source:
          $ref: "#/components/schemas/DataMode"
        message:
          type: string
        eventsIndexed:
          type: integer
        signaturesScanned:
          type: integer
        unknownTransactions:
          type: integer
        lastIndexedAt:
          type: string
          format: date-time
          nullable: true
        simulatedTrades:
          type: integer
        warnings:
          type: array
          items:
            type: string
      required: [address, displayAddress, isDemo, demoLabel, state, source, message, eventsIndexed, signaturesScanned, unknownTransactions, lastIndexedAt, simulatedTrades, warnings]
...
    Portfolio:
      type: object
      properties:
        address:
          type: string
        displayAddress:
          type: string
        isDemo:
          type: boolean
        asOf:
          type: string
          format: date-time
        method:
          $ref: "#/components/schemas/CostMethod"
        currency:
          type: string
        totals:
          $ref: "#/components/schemas/PortfolioTotals"
        positions:
          type: array
          items:
            $ref: "#/components/schemas/Position"
        allocation:
          type: array
          items:
            $ref: "#/components/schemas/IssuerAllocation"
        pricing:
          $ref: "#/components/schemas/PricingStatus"
        walletStatus:
          $ref: "#/components/schemas/WalletStatus"
        assumptions:
          type: array
          items:
            type: string
      required: [address, displayAddress, isDemo, asOf, method, currency, totals, positions, allocation, pricing, walletStatus, assumptions]
    Lot:
      type: object
      properties:
        id:
          type: string
        mint:
          type: string
        symbol:
          type: string
        issuer:
          $ref: "#/components/schemas/IssuerId"
        openedAt:
          type: string
          format: date-time
        openSignature:
          type: string
          nullable: true
        openKind:
          type: string
          enum: [buy, transfer_in, wrapper_swap, opening_balance]
        quantity:
          type: number
          description: Original quantity of exposure at acquisition
        remainingQuantity:
          type: number
        costPerShare:
          type: number
          nullable: true
        costBasis:
          type: number
          nullable: true
        remainingCostBasis:
          type: number
          nullable: true
        marketValue:
          type: number
          nullable: true
        unrealizedPnl:
          type: number
          nullable: true
        realizedPnl:
          type: number
        holdingDays:
          type: integer
        term:
          type: string
          enum: [short, long]
        status:
          type: string
          enum: [open, partial, closed]
        basisStatus:
          type: string
          enum: [complete, estimated, unknown]
        basisNote:
          type: string
          nullable: true
        closedAt:
          type: string
          format: date-time
          nullable: true
        explorerUrl:
          type: string
          nullable: true
      required: [id, mint, symbol, issuer, openedAt, openSignature, openKind, quantity, remainingQuantity, costPerShare, costBasis, remainingCostBasis, marketValue, unrealizedPnl, realizedPnl, holdingDays, term, status, basisStatus, basisNote, closedAt, explorerUrl]
...
    ActivityPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/LedgerEvent"
        nextCursor:
          type: string
          nullable: true
        total:
          type: integer
      required: [items, nextCursor, total]
    CorporateActionEvent:
      type: object
      properties:
        id:
          type: string
        mint:
          type: string
        symbol:
          type: string
        issuer:
          $ref: "#/components/schemas/IssuerId"
        kind:
          type: string
          enum: [dividend_reinvested, stock_split, reverse_split, multiplier_change, pending_multiplier]
        kindLabel:
          type: string
        previousMultiplier:
          type: number
        newMultiplier:
          type: number
        ratio:
          type: number
        effectiveAt:
          type: string
          format: date-time
        detectedAt:
          type: string
          format: date-time
        quantityBefore:
          type: number
          nullable: true
        quantityAfter:
          type: number
          nullable: true
        valueEffect:
          type: number
          nullable: true
        confidence:
          type: string
          enum: [confirmed, inferred, scripted]
        source:
          type: string
        note:
          type: string
        explorerUrl:
          type: string
          nullable: true
      required: [id, mint, symbol, issuer, kind, kindLabel, previousMultiplier, newMultiplier, ratio, effectiveAt, detectedAt, quantityBefore, quantityAfter, valueEffect, confidence, source, note, explorerUrl]
...
    PricingStatus:
      type: object
      properties:
        provider:
          type: string
          enum: [pyth, jupiter, prestocks, mixed, demo, none]
        providerLabel:
          type: string
        mode:
          $ref: "#/components/schemas/DataMode"
        headline:
          type: string
        detail:
          type: string
        asOf:
          type: string
          format: date-time
        pythConfigured:
          type: boolean
        pythAuthorized:
          type: boolean
          nullable: true
        feedsResolved:
          type: integer
        feedsTotal:
          type: integer
        usSession:
          $ref: "#/components/schemas/MarketSession"
        feeds:
          type: array
          items:
            $ref: "#/components/schemas/PricingFeedStatus"
      required: [provider, providerLabel, mode, headline, detail, asOf, pythConfigured, pythAuthorized, feedsResolved, feedsTotal, usSession, feeds]
    StatementRequest:
      type: object
      properties:
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        method:
          $ref: "#/components/schemas/CostMethod"
        title:
          type: string
      required: [periodStart, periodEnd]
    StatementSummary:
      type: object
      properties:
        id:
          type: string
        address:
          type: string
        title:
          type: string
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        method:
          $ref: "#/components/schemas/CostMethod"
        generatedAt:
          type: string
          format: date-time
        hash:
          type: string
        closingValue:
          type: number
        realizedPnl:
          type: number
        proofStatus:
          type: string
          enum: [none, pending, confirmed, failed, simulated]
      required: [id, address, title, periodStart, periodEnd, method, generatedAt, hash, closingValue, realizedPnl, proofStatus]
...
    StatementProof:
      type: object
      properties:
        status:
          type: string
          enum: [none, pending, confirmed, failed, simulated]
        kind:
          type: string
          enum: [onchain_memo, simulated, none]
        hash:
          type: string
        memo:
          type: string
        signature:
          type: string
          nullable: true
        slot:
          type: integer
          nullable: true
        signer:
          type: string
          nullable: true
        confirmedAt:
          type: string
          format: date-time
          nullable: true
        explorerUrl:
          type: string
          nullable: true
        message:
          type: string
      required: [status, kind, hash, memo, signature, slot, signer, confirmedAt, explorerUrl, message]
    Statement:
      type: object
      properties:
        id:
          type: string
        address:
          type: string
        displayAddress:
          type: string
        isDemo:
          type: boolean
        title:
          type: string
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        method:
          $ref: "#/components/schemas/CostMethod"
        generatedAt:
          type: string
          format: date-time
        currency:
          type: string
        hash:
          type: string
        hashAlgorithm:
          type: string
        totals:
          $ref: "#/components/schemas/StatementTotals"
        positions:
          type: array
          items:
            $ref: "#/components/schemas/StatementPosition"
        activity:
          type: array
          items:
            $ref: "#/components/schemas/LedgerEvent"
        corporateActions:
          type: array
          items:
            $ref: "#/components/schemas/CorporateActionEvent"
        closedLots:
          type: array
          items:
            $ref: "#/components/schemas/LotRelief"
        pricing:
          $ref: "#/components/schemas/PricingStatus"
        assumptions:
          type: array
          items:
            type: string
        dataSources:
          type: array
          items:
            $ref: "#/components/schemas/SourceStatus"
        proof:
          $ref: "#/components/schemas/StatementProof"
        csvUrl:
          type: string
        pdfUrl:
          type: string
      required: [id, address, displayAddress, isDemo, title, periodStart, periodEnd, method, generatedAt, currency, hash, hashAlgorithm, totals, positions, activity, corporateActions, closedLots, pricing, assumptions, dataSources, proof, csvUrl, pdfUrl]
    NotarizationPayload:
      type: object
      properties:
        statementId:
          type: string
        hash:
          type: string
        memo:
          type: string
          description: Exact UTF-8 memo the wallet must sign
        memoProgramId:
          type: string
        cluster:
          type: string
        mode:
          type: string
          enum: [onchain, simulated]
        instructions:
          type: string
        transaction:
          type: string
          nullable: true
          description: Base64 encoded unsigned versioned transaction containing the memo, present when payer was supplied and mode is onchain
        lastValidBlockHeight:
          type: integer
          nullable: true
        proof:
          $ref: "#/components/schemas/StatementProof"
      required: [statementId, hash, memo, memoProgramId, cluster, mode, instructions, transaction, lastValidBlockHeight, proof]
    NotarizationSubmission:
      type: object
      properties:
        signature:
          type: string
          nullable: true
          description: Signature of the memo transaction sent by the wallet
        simulate:
          type: boolean
          description: Create a clearly labelled simulated proof instead of verifying on chain
      required: [simulate]
    TradeQuoteRequest:
      type: object
      properties:
        mint:
          type: string
        quantity:
          type: number
          description: Exposure quantity to sell, after multiplier
        method:
          $ref: "#/components/schemas/CostMethod"
        slippageBps:
          type: integer
        outputMint:
          type: string
          description: Defaults to USDC
      required: [mint, quantity]
...
    TradePrepareRequest:
      type: object
      properties:
        quoteId:
          type: string
        userPublicKey:
          type: string
      required: [quoteId, userPublicKey]
    TradeTransaction:
      type: object
      properties:
        quoteId:
          type: string
        swapTransaction:
          type: string
          description: Base64 encoded versioned transaction to sign
        lastValidBlockHeight:
          type: integer
        rpcUrl:
          type: string
        expiresAt:
          type: string
          format: date-time
      required: [quoteId, swapTransaction, lastValidBlockHeight, rpcUrl, expiresAt]
...
    TradeSimulateRequest:
      type: object
      properties:
        quoteId:
          type: string
      required: [quoteId]
...